Create query_planning.md
This multi-step query thing was not as simple as previously thought. This approach seems to be pretty capable. Downside is that each query now takes atleast 2. Some later fact requests like summarization could be called with curie for cheaper prices.
I also tried to create pseudo-chat from this back-and-forth, but that was too hard to keep well formed.